Rename from ts to js
