+++ src/util/railsAxios.js
@@ -1,0 +1,21 @@
+import axios from 'axios';
+import getEnv from './environment';
+
+const railsAxios = (jwt) => {
+  const JSON_HEADERS = {
+    Accept: 'application/json',
+    'Content-Type': 'application/json',
+  };
+  const AUTH_HEADER = {
+    Authorization: `Bearer ${jwt}`,
+  };
+  const { API_BASE_URL } = getEnv();
+  return axios.create({
+    headers: jwt
+      ? { ...JSON_HEADERS, ...AUTH_HEADER }
+      : { ...JSON_HEADERS },
+    baseURL: API_BASE_URL,
+  });
+};
+
+export default railsAxios;
